Apply homepage logic #23
Besides creating the activity, the flow of the app with this new
activity has been established.
