+++ app/src/main/java/com/letslunch/agileteam8/letslunch/ReminderAlarmService.java
@@ -27,7 +27,7 @@ public class ReminderAlarmService extends IntentService {
         builder.setContentTitle("Lets Lunch reminder");
         builder.setContentText("What's your lunch plan?");
         builder.setSmallIcon(R.mipmap.ic_launcher);
-        Intent notifyIntent = new Intent(this, MainActivity.class);
+        Intent notifyIntent = new Intent(this, homePage.class);
         PendingIntent pendingIntent = PendingIntent.getActivity(this, 2, notifyIntent, PendingIntent.FLAG_UPDATE_CURRENT);
         //to be able to launch your activity from the notification
         builder.setContentIntent(pendingIntent);
